+++ myems-web/src/components/MyEMS/Microgrid/MicrogridReporting.js
@@ -34,8 +34,10 @@ import { endOfDay } from 'date-fns';
 import AppContext from '../../../context/Context';
 import { Link, useLocation } from 'react-router-dom';
 import DateRangePickerWrapper from '../common/DateRangePickerWrapper';
-const DetailedDataTable = loadable(() => import('../common/DetailedDataTable'));
+import blankPage from '../../../assets/img/generic/blank-page.png';
 
+
+const DetailedDataTable = loadable(() => import('../common/DetailedDataTable'));
 
 const MicrogridReportingEnergy = ({ setRedirect, setRedirectUrl, t }) => {
   let current_moment = moment();
@@ -939,7 +941,10 @@ const MicrogridReportingEnergy = ({ setRedirect, setRedirectUrl, t }) => {
           </Form>
         </CardBody>
       </Card>
-      <div style={{visibility: resultDataHidden ? 'hidden' : 'visible'}}>
+      <div  style={{ visibility: resultDataHidden ? 'visible' : 'hidden', display: resultDataHidden ? '': 'none' }}>
+          <img className="img-fluid" src={blankPage} alt="" />
+      </div>
+      <div style={{ visibility: resultDataHidden ? 'hidden' : 'visible', display: resultDataHidden ? 'none': ''  }}>
         {/* <div className="card-deck" >
           {cardSummaryList.map(cardSummaryItem => (
             <CardSummary